GenerateControllerInterfaces and GenerateAbstractController settings are deleted. ControllerGenerationFormat parameter is added. (default:partial)Test classes are deleted and new test class is added.
