+++ src/NSwag.CodeGeneration.CSharp.Tests/NSwag.CodeGeneration.CSharp.Tests.csproj
@@ -63,11 +63,10 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ArrayParameterTests.cs" />
+    <Compile Include="ControllerGenerationFormatTests.cs" />
     <Compile Include="CSharpClientSettingsTests.cs" />
     <Compile Include="FileDownloadTests.cs" />
     <Compile Include="FormParameterTests.cs" />
-    <Compile Include="GenerateAbstractControllersTests.cs" />
-    <Compile Include="GenerateControllerInterfacesTests.cs" />
     <Compile Include="HeadRequestTests.cs" />
     <Compile Include="OptionalParameterTests.cs" />
     <Compile Include="ParameterTests.cs" />
