+++ src/NSwag.CodeGeneration.CSharp/Models/CSharpControllerTemplateModel.cs
@@ -65,9 +65,7 @@ namespace NSwag.CodeGeneration.CSharp.Models
         /// <summary>Gets a value indicating whether to generate optional parameters.</summary>
         public bool GenerateOptionalParameters => _settings.GenerateOptionalParameters;
 
-        /// <summary>Gets a value indicating whether to generate controller interfaces.</summary>
-        public bool GenerateControllerInterfaces => _settings.GenerateControllerInterfaces;
-        /// <summary>Gets a value indicating whether to generate abstract controllers.</summary>
-        public bool GenerateAbstractControllers => _settings.GenerateAbstractControllers;
+        /// <summary>Gets a value indicating whether to controller generation format.(default: partial;abstract, partial.)</summary>
+        public string ControllerGenerationFormat => _settings.ControllerGenerationFormat;
     }
 }
