+++ src/NSwag.CodeGeneration.CSharp/SwaggerToCSharpGeneratorSettings.cs
@@ -28,8 +28,7 @@ namespace NSwag.CodeGeneration.CSharp
 
             GenerateResponseClasses = true;
             ResponseClass = "SwaggerResponse";
-            GenerateControllerInterfaces = true;
-            GenerateAbstractControllers = false;
+            ControllerGenerationFormat = "partial";
         }
 
         /// <summary>Gets or sets the CSharp generator settings.</summary>
@@ -52,9 +51,9 @@ namespace NSwag.CodeGeneration.CSharp
 
         /// <summary>Gets or sets the name of the response class (supports the '{controller}' placeholder).</summary>
         public string ResponseClass { get; set; }
-        /// <summary>Gets a value indicating whether to generate controller interfaces.</summary>
-        public bool GenerateControllerInterfaces { get; set; }
-        /// <summary>Gets a value indicating whether to generate abstract controllers.</summary>
-        public bool GenerateAbstractControllers { get; set; }
+
+        /// <summary>Gets a value indicating whether to controller generation format.(default: partial;abstract, partial.)</summary>
+        public string ControllerGenerationFormat { get; set; }
+  
     }
 }
