+++ src/NSwag.CodeGeneration.CSharp/Templates/ControllerTemplate.tt
@@ -1,9 +1,12 @@
 ﻿<#@ template visibility="internal" #>
 <#@ import namespace="NJsonSchema" #>
 
+<#
+    var isPartial = Model.ControllerGenerationFormat == "partial";
+    var isAbstract = Model.ControllerGenerationFormat == "abstract";
+#>
 
-<# if (Model.GenerateControllerInterfaces){#>
-[System.CodeDom.Compiler.GeneratedCode("NSwag", "<#=SwaggerDocument.ToolchainVersion#>")]
+<# if (isPartial){#>[System.CodeDom.Compiler.GeneratedCode("NSwag", "<#=SwaggerDocument.ToolchainVersion#>")]
 public interface I<#=Model.Class#>Controller
 {
 <#foreach(var operation in Model.Operations){#>
@@ -16,31 +19,30 @@ public interface I<#=Model.Class#>Controller
 
 <#}#>
 }
-<# }#>
-
+<#}#>
 [System.CodeDom.Compiler.GeneratedCode("NSwag", "<#=SwaggerDocument.ToolchainVersion#>")]
 <#if(Model.HasBasePath){#>
 [System.Web.Http.RoutePrefix("<#=Model.BasePath#>")]
 <#}#>
-public <# if(Model.GenerateAbstractControllers){#>abstract class <#=Model.Class#>ControllerBase : <#}#><#else{#>partial class <#=Model.Class#>Controller :  <#}#><#if(Model.HasBaseClass){#><#=Model.BaseClass#><#}else{#>System.Web.Http.ApiController<#}#> 
-{<# if (Model.GenerateControllerInterfaces){#>    
-	private I<#=Model.Class #>Controller _implementation; 
+<# if (isPartial){#>
+public partial class <#=Model.Class#>Controller : <#if(Model.HasBaseClass){#><#=Model.BaseClass#><#}else{#>System.Web.Http.ApiController<#}#> 
+{
+    private I<#=Model.Class#>Controller _implementation; 
 
-    public <#=Model.Class #>Controller(I<#=Model.Class #>Controller implementation)
+    public <#=Model.Class#>Controller(I<#=Model.Class#>Controller implementation)
     {
         _implementation = implementation; 
     }
-<#}#>
-	<#foreach(var operation in Model.Operations){#>
+
+<#foreach(var operation in Model.Operations){#>
 <#  if(operation.HasSummary){#>    /// <summary><#=ConversionUtilities.ConvertCSharpDocBreaks(operation.Summary, 1)#></summary>
 <#  }foreach (var parameter in operation.Parameters){#>
 <#      if(parameter.HasDescription){#>    /// <param name="<#=parameter.VariableName#>"><#=ConversionUtilities.ConvertCSharpDocBreaks(parameter.Description, 1)#></param>
 <#  }}if(operation.HasResultDescription){#>    /// <returns><#=ConversionUtilities.ConvertCSharpDocBreaks(operation.ResultDescription, 1)#></returns>
 <#  }if(operation.IsDeprecated){#>    [System.Obsolete]
-<#}#>	[System.Web.Http.Http<#=operation.HttpMethodUpper#>, System.Web.Http.Route("<#=operation.Path#>")]
-	public <# if(Model.GenerateAbstractControllers){#>abstract <#}#> <# if(Model.WrapResponses){#><# if(!Model.GenerateAbstractControllers){#>async <#}#>System.Threading.Tasks.Task<System.Net.Http.HttpResponseMessage><#}else{#><#=operation.ResultType#><#}#> <#=operation.ActualOperationName#>(<#foreach(var parameter in operation.Parameters){#><#=parameter.Type#> <#=parameter.VariableName#><# if(Model.GenerateOptionalParameters && parameter.IsOptional){#> = null<#}#><# if(!parameter.IsLast){#>, <#}}#>)<# if(Model.GenerateAbstractControllers){#>;
-<#}else {#>	{<#}#>
-<# if(Model.GenerateControllerInterfaces){#>
+<#  }#>    [System.Web.Http.Http<#=operation.HttpMethodUpper#>, System.Web.Http.Route("<#=operation.Path#>")]
+    public <#if(Model.WrapResponses){#>async System.Threading.Tasks.Task<System.Net.Http.HttpResponseMessage><#}else{#><#=operation.ResultType#><#}#> <#=operation.ActualOperationName#>(<#foreach(var parameter in operation.Parameters){#><#=parameter.Type#> <#=parameter.VariableName#><#if(Model.GenerateOptionalParameters && parameter.IsOptional){#> = null<#}#><#if(!parameter.IsLast){#>, <#}}#>)
+    {    
 <#          if (Model.WrapResponses){#>
         var result = await _implementation.<#=operation.ActualOperationName#>Async(<#foreach(var parameter in operation.Parameters){#><#=parameter.VariableName#><#if(!parameter.IsLast){#>, <#}}#>);
 
@@ -52,24 +54,29 @@ public <# if(Model.GenerateAbstractControllers){#>abstract class <#=Model.Class#
 
         return response;
 <#          }else{#>
-
-		return _implementation.<#=operation.ActualOperationName#>Async(<#foreach(var parameter in operation.Parameters){#><#=parameter.VariableName#><#if(!parameter.IsLast){#>, <#}}#>);
+        return _implementation.<#=operation.ActualOperationName#>Async(<#foreach(var parameter in operation.Parameters){#><#=parameter.VariableName#><#if(!parameter.IsLast){#>, <#}}#>);
 <#          }#>
+        
+    }
 
-    }<#}else {#>
-<# if (!Model.GenerateAbstractControllers){#>
-<#          if (Model.WrapResponses){#>
-        var result = System.Threading.Tasks.Task<<#=operation.UnwrappedResultType#>>.FromResult(new <#=operation.UnwrappedResultType#>());
-
-        var status = System.Net.HttpStatusCode.OK;
-        HttpResponseMessage response = Request.CreateResponse(status<# if(operation.UnwrappedResultType != "void"){#>, result<#}#>);
-    
-        return response;
-<#          }else{#>
-
-        return System.Threading.Tasks.Task<<#=operation.UnwrappedResultType#>>.FromResult(new <#=operation.UnwrappedResultType#>());
-	<#}#>}
-	<#}#>
-<#}#>
 <#}#>
 }
+<# }else if (isAbstract){#>
+
+public abstract class <#=Model.Class#>ControllerBase : <#if(Model.HasBaseClass){#><#=Model.BaseClass#><#}else{#>System.Web.Http.ApiController<#}#>
+
+{
+<#foreach(var operation in Model.Operations){#>
+<#  if(operation.HasSummary){#>    /// <summary><#=ConversionUtilities.ConvertCSharpDocBreaks(operation.Summary, 1)#></summary>
+<#  }foreach (var parameter in operation.Parameters){#>
+<#      if(parameter.HasDescription){#>    /// <param name="<#=parameter.VariableName#>"><#=ConversionUtilities.ConvertCSharpDocBreaks(parameter.Description, 1)#></param>
+<#  }}if(operation.HasResultDescription){#>    /// <returns><#=ConversionUtilities.ConvertCSharpDocBreaks(operation.ResultDescription, 1)#></returns>
+<#  }if(operation.IsDeprecated){#>    [System.Obsolete]
+<#  }#>    [System.Web.Http.Http<#=operation.HttpMethodUpper#>, System.Web.Http.Route("<#=operation.Path#>")]
+	public abstract <# if(Model.WrapResponses){#>System.Threading.Tasks.Task<System.Net.Http.HttpResponseMessage><#}else{#><#=operation.ResultType#><#}#> <#=operation.ActualOperationName#>(<#foreach(var parameter in operation.Parameters){#><#=parameter.Type#> <#=parameter.VariableName#><# if(Model.GenerateOptionalParameters && parameter.IsOptional){#> = null<#}#><# if(!parameter.IsLast){#>, <#}}#>);
+
+<#}#>
+}
+<#}#>
+
+
